integration tests: move transcription test and add basic router test
