+++ Cargo.toml
@@ -9,6 +9,7 @@ axum-macros = "0.5.0"
 derive_builder = "0.20.2"
 derive_more = { version = "2.0.1", features = ["from"] }
 dotenvy = "0.15.7"
+http-body-util = "0.1.3"
 mockall = "0.13.1"
 regex = "1.10.6"
 reqwest = { version = "0.12.7", features = ["json"] }
@@ -16,13 +17,13 @@ serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.141"
 sqlx = { version = "0.8.6", features = ["runtime-tokio", "postgres"] }
 tokio = { version = "1.35.1", features = ["full", "time", "rt"] }
+tower = "0.5.2"
 tower-http = { version = "0.6.6", features = ["fs", "trace"] }
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 url = "2.5.4"
 
 [dev-dependencies]
-anyhow = "1.0.98"
 httpc-test = "0.1.1"
 
 [lints.rust]
