+++ src/web/services/transcript.rs
@@ -85,7 +85,7 @@ pub async fn summarize_by_url(url: &YTUrl, pool: &PgPool) -> Result<String> {
 }
 
 /// remove timestamps and duplicate lines
-pub fn clean_vtt(transcript: &str) -> String {
+fn clean_vtt(transcript: &str) -> String {
 	let mut lines = transcript.lines();
 	// skip header
 	lines.find(|l| l.starts_with("Language"));
@@ -107,77 +107,4 @@ pub fn clean_vtt(transcript: &str) -> String {
 		.filter(|l| !l.is_empty())
 		.collect::<Vec<_>>()
 		.join(" ")
-}
-
-#[cfg(test)]
-mod tests {
-	use anyhow::Result;
-	use mockall::predicate;
-
-	use crate::web::services::youtube::MockYtServiceTrait;
-
-	use super::*;
-
-	const VTT: &str =  "WEBVTT
-Kind: captions
-Language: en
-00:00:00.580 --> 00:00:01.910 align:start position:0%
-[Music]
-00:00:01.910 --> 00:00:01.920 align:start position:0%
-[Music]
-
-00:00:01.920 --> 00:00:04.150 align:start position:0%
-[Music]
-you<00:00:02.040><c> know</c><00:00:02.200><c> what's</c><00:00:02.520><c> really</c><00:00:02.840><c> not</c><00:00:03.120><c> fun</c><00:00:03.679><c> recording</c>
-
-00:00:04.150 --> 00:00:04.160 align:start position:0%
-you know what's really not fun recording
-
-00:00:04.160 --> 00:00:06.510 align:start position:0%
-you know what's really not fun recording
-an<00:00:04.359><c> entire</c><00:00:04.880><c> video</c><00:00:05.160><c> for</c><00:00:05.400><c> 30</c><00:00:05.720><c> minutes</c><00:00:06.200><c> and</c><00:00:06.319><c> then</c>
-
-00:00:06.510 --> 00:00:06.520 align:start position:0%
-an entire video for 30 minutes and then
- 
-
-00:00:06.520 --> 00:00:08.669 align:start position:0%
-an entire video for 30 minutes and then
-realizing<00:00:07.359><c> you</c><00:00:07.520><c> forgot</c><00:00:07.839><c> to</c><00:00:08.080><c> plug</c><00:00:08.280><c> in</c><00:00:08.440><c> your</c>";
-
-	const CLEAN_VTT: &str =
-		"[Music] you know what's really not fun recording an entire video for 30 minutes and then";
-
-	#[test]
-	fn should_convert_vtt_to_text() {
-		assert_eq!(clean_vtt(VTT), CLEAN_VTT);
-	}
-
-	#[sqlx::test]
-	async fn should_get_and_cache_transcript(pool: PgPool) -> Result<()> {
-		let url = YTUrl::try_from(
-			"https://www.youtube.com/watch?v=DjcC6p_8fpE&pp=ygUWamFjb2IgYXNwZXIgdHlwZXNjcmlwdA%3D%3D",
-		)?;
-
-		let mut yt_service = MockYtServiceTrait::new();
-		yt_service
-			.expect_fetch_captions()
-			.with(predicate::eq(url.clone()))
-			.times(1)
-			.returning(|_url| Ok(VTT.to_owned()));
-
-		let transcript = get_transcript_by_url(&url, &pool, yt_service).await?;
-		assert_eq!(transcript, CLEAN_VTT);
-
-		// should be stored in DB
-		let mut yt_service = MockYtServiceTrait::new();
-		yt_service
-			.expect_fetch_captions()
-			.times(0);
-
-		let transcript = get_transcript_by_url(&url, &pool, yt_service).await?;
-		assert_eq!(transcript, CLEAN_VTT);
-
-		Ok(())
-	}
 }
